Modified app to use router-outlet. Added new API services. Improved event stacking function
App was modified to use <router-outlet> so that the app can be extended.

Added new API services to test interfaces with the backend API.

There was a minor issue with multi-day event stacking that made events overlay in some situations. This was fixed looking for existing positions before adding new positions.